+++ src/main/java/com/MA/AlrightBet/Service/User_impl.java
@@ -1,11 +1,7 @@
 package com.MA.AlrightBet.Service;
 
 
-import com.MA.AlrightBet.Dao.AdminDao;
-import com.MA.AlrightBet.Dao.BetDao;
-import com.MA.AlrightBet.Dao.FightCardDao;
 import com.MA.AlrightBet.Dao.UserDao;
-import com.MA.AlrightBet.Entity.FightCard;
 import com.MA.AlrightBet.Entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +11,6 @@ import java.util.Optional;
 
 @Service
 public class User_impl implements UserService {
-
 
 
     @Autowired
@@ -36,15 +31,19 @@ public class User_impl implements UserService {
         return user.get();
     }
 
-
     @Override
     public User create_user(User user) {
         return this.userDao.save(user);
     }
 
     @Override
-    public void delete_user(int id) {
-        this.userDao.deleteById(id);
+    public Boolean delete_user(int id) {
+        Optional<User> q = this.userDao.findById(id);
+        if (q.isPresent()) {
+            this.userDao.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
